upload_files_project/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt
djoser
drf-yasg
psycopg2-binary
celery[redis]
Pillow
gunicorn
flower